1.5 compatibility

+++ SDVModTest/UIElements/ExperiencePointDisplay.cs
@@ -1,10 +1,5 @@
 ﻿using Microsoft.Xna.Framework;
 using StardewValley;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace UIInfoSuite.UIElements
 {
@@ -26,8 +21,8 @@ namespace UIInfoSuite.UIElements
             --_alpha;
             Game1.drawWithBorder(
                 "Exp " + _experiencePoints,
-                Color.DarkSlateGray * ((float)_alpha / 100f),
-                Color.PaleTurquoise * ((float)_alpha / 100f),
+                Color.DarkSlateGray * (_alpha / 100f),
+                Color.PaleTurquoise * (_alpha / 100f),
                 new Vector2(_position.X - 28, _position.Y - 130),
                 0.0f,
                 0.8f,
